(GH-1836) Add PuppetObject interface to Bolt data types
This adds the `Puppet::Pops::Types::PuppetObject` interface to Bolt data
types and their associated classes. Previously, Bolt's data types were
missing this interface, which could lead to Puppet attempting to access
already-deserialized objects with the `[]` operator during the
deserialization process when local references are present. Since Bolt's
implementation classes do not implement the `[]` operator, and doing so
is undesirable, adding the `PuppetObject` interface is needed for
deserialization.

!bug

* **Add `PuppetObject` interface to Bolt data types**
  ([#1836](https://github.com/puppetlabs/bolt/issues/1836))

  Bolt data types would sometimes not be deserialized correctly when
  using `apply` blocks in plans. All Bolt data types now implement the
  `PuppetObject` interface so they can be deserialized correctly.