(GH-1829) Use system project directory if homedir isn't defined
Bolt is sometimes run in processes where no homedir is defined, causing
the expansion of `~` to fail. This was previously taken care of in most
places in #1676 but missed the default project directory at
`~/.puppetlabs/bolt`. Now when we try to load the path if it fails it
will fall back to the system-wide configuration path at
`/etc/puppetlabs/bolt` as the default project directory.

!bug

* **Fall back to system-wide config path if homedir expansion fails**

  We now fall back to `/etc/puppetlabs/bolt` as the default project
  directory if expanding the homedir fails.