(maint) Jruby compatability with default Bolt::Config
The Jruby environment used in Orchestrator does not have the `Etc::SC_OPEN_MAX` constant defined. Given the default is evaluated even if the config is overridden this commit updates the logic to decide whether to attempt to read the system config. This change also caches the result for the `default_concurrency` method so that the system config is not fetched multiple times.

!bug

* **Do not attempt to use `Etc::SC_OPEN_MAX` when it is not defined** ([1858](https://github.com/puppetlabs/bolt/pull/1858))

  When the `SC_OPEN_MAX` constant is not defined (for example when running under JRuby) do not attempt to use it to determine default concurrency.