(GH-1847) Use task parameter default when parameter is `Undef`
This updates the `run_task` function to use a task parameter default
when a parameter is specified but has a value of `Undef`. Previously, if
a parameter was specified with a value of `Undef`, the default would not
be used and the `Undef` value would be passed to the task.

!feature

* **Use task parameter default when specified parameter is `Undef`**
  ([#1847](https://github.com/puppetlabs/issues/1847))

  Task parameters that are specified with a value of `Undef` will now
  use the default parameter value if one is defined in the task's
  metadata.