(GH-1874) Add resource plan function to lookup resource on a target
This adds a `resource()` plan function that takes a target, type, and
title and returns either the resource object or nil. The function is a
plan function, so either `$target.resource('type', 'title')` or
`resource($target, 'type', 'title')` can be used. It looks up the
resource based on the computed reference in the target's resources hash,
and return nil if the resource isn't found.

Closes #1874

!feature

* **Add resource plan function to lookup a resource on a target** ([#1874](https://github.com/puppetlabs/bolt/issues/1874))

  This adds a `resource` plan function that can be used to lookup a
  resource on a Target object by 'type' and 'title'.