(GH-1872) Accept `resource_type` key in `set_resources` plan function
This updates the `set_resources` plan function to accept a
`resource_type` key as part of the resource data hash, in addition to
`type`.

When creating a `ResourceInstance` object from a hash, a `type` key is
expected. However, when using resource data hashes returned from an
apply block, the resource's type is set under the `resource_type` key.
This made it impossible to directly use this hash to create a
`ResourceInstance` object, as the hash did not have the required `type`
key. The `set_resources` function now falls back to using the
`resource_type` key if the `type` key is not present.

It's important to note that calling `ResourceInstance.new` with a hash
that has a `resource_type` key instead of a `type` key will still raise
an exception.

To keep the behavior of `ResourceInstance.new` predictable, this also
removes the ability to use a `parameters` key instead of a `state` key
in the implementation class.

!feature

* **Accept `resource_type` key in resource data hash for `set_resources`
  plan function**
  ([#1872](https://github.com/puppetlabs/bolt/issues/1872))

  The `set_resources` function now accepts resource data hashes that
  have a `resource_type` key instead of a `type` key. This allows users
  to set resources directly from reports from an apply block, which set
  a resource's type under the `resource_type` key.