(GH-1862) Specify preferred  algorithms for SSH transport connections
This adds the following SSH transport configuration options:

- `encryption-algorithms`
- `host-key-algorithms`
- `kex-algorithms`
- `mac-algorithms`

Each of these options accepts an array of preferred algorithms to use when
establishing a connection using the SSH transport. When a `*-algorithms`
option is set, `net-ssh` is configured to only attempt to establish a
connection to the target using algorithms in the list.

Each `*-algorithms` option accepts a special `defaults` item, which will
automatically add all of default, supported algorithms for `net-ssh` to
the list. For example, the following config:

```yaml
ssh:
  encryption-algorithms:
    - defaults
```

Would expand to the following:

```yaml
ssh:
  encryption-algorithms:
    - aes-256-ctr
    - aes-192-ctr
    - aes-128-ctr
```

Starting with the release of `net-ssh` 6.0, several algorithms were
deprecated and disabled by default. This prevented users from using
these algorithms when establishing connections with targets. These
options allow users to enable these deprecated algorithms.

!feature

* **Specify preferred algorithms for SSH transport connections**
  ([#1862](https://github.com/puppetlabs/bolt/issues/1862))

  Users can now specify a list of preferred algorithms to use when
  establishing connections with targets using the SSH transport with the
  `encryption-algorithms`, `host-key-algorithms`, `kex-algorithms`, and
  `mac-algorithms` config options. Each option accepts an array of
  algorithms and overrides the default list of preferred algorithms.
  You can read more about these options in the [Bolt configuration
  reference](https://puppet.com/docs/bolt/latest/bolt_configuration_reference.html#ssh).