(GH-1870) Fix low ulimit warning
The low ulimit warning had a few issues:
1. It was displayed whether or not the concurrency limit was reached
   during the Bolt lifecycle
2. It miscalculated the concurrency at which users might cause the
   EMFILE error
3. It suggested setting concurrency on the CLI instead of in config

This PR rectifies each of those issues. We now only display the warning
once per Bolt run if a particular step causes the lowered concurrency
limit to be reached. The warning will only be displayed if the ulimit is
too low and the user hasn't configured the concurrency. This also sets
the default concurrency to `ulimit/7` to more accurately estimated how
many file descriptors each thread uses and try to prevent errors. Lastly
it updates the (now unprecedentedly long) warning to explain the issue,
suggest configuring ulimit or concurrency, and link to the known issues
document.

Closes #1870

!bug

* **Improve low ulimit warning** ([1870](https://github.com/puppetlabs/bolt/issues/1870))

  Users running Bolt with a low ulimit should only be warned if the
  number of targets they're running against may cause file limit issues.