Do not check if unrecognized plan parameters are sensitive
This updates the `run_plan` function to not check if a plan parameter
has a `Sensitive` type if the parameter is not recognized by the plan.
Previously, if a parameter was passed to a plan that the plan was not
expecting, the `run_plan` function would attempt to check the
parameter's type expression to determine whether it should be wrapped as
`Sensitive`. Since the parameter is unrecognized, it would raise an
error when attempting to access its type expression.

!bug

* **Raise correct error when passing unknown plan parameters**
  ([#1886](https://github.com/puppetlabs/bolt/pull/1886))

  Bolt was raising an obscure error when a plan received an unknown
  parameter. It now raises the correct error indicating that the
  parameter is unknown.