(GH-1888) Create separate reference docs for config files
This adds separate reference docs for each of Bolt's configuration
files. Previously, configuration options were listed under a single
monolithic 'Bolt configuration options' page, which made it difficult to
know which options were available to which files. Reference pages are
now available for the following files:

- `bolt.yaml`
- `inventory.yaml`
- `bolt-project.yaml`
- `bolt-defaults.yaml`

This also adds a separate reference page for transport configuration
options, as these options can be set in multiple files.

This updates the transport config classes to use a single
`TRANSPORT_OPTIONS` hash to pull available options data from. Instead of
each transport class defining the option data separately (leading to
duplicate data across the classes), each class now contains a list of
option names that are available to it. When pulling the option data for
validation, the transport class will slice the appropriate options from
the `TRANSPORT_OPTIONS` hash.

This also updates the `assert_type` method on the base transport config
class to accept an array of types. This allows for a handful of config
options that accept multiple types (such as `private-key` which accepts
either a String or a Hash) to document as such and perform automatic
validation.

!no-release-note