(GH-733) Document connecting with configuration not supported by net-ssh
This adds a section to the experimental features page under external SSH
about how to use the external SSH configuration option to enable SSH
configuration that isn't supported by the Ruby net-ssh library. It
includes an example of both defining the command to include the
configuration, and an example with configuration defined in
`~/.ssh/config`. It also removes the OpenSSH 7.8 key format known issue,
as it is no longer an issue since we ship Bolt with the gems to support
the new format.

Closes #733

!no-release-note