(GH-1867) Document how privilege escalation works in Bolt
This adds a net new page to the docs titled `Escalating privilege with
Bolt` which details how Bolt builds the command to change users, how to
configure that command, and what limitations Bolt has with respect to
escalating privilege.

The page also includes a list of escalation-related configuration
options and command line flags, which are generated from Bolt itself the
same way the command line reference and configuration reference pages
are generated. The generation of the options on that page is run with a
new rake task `docs:privilege_escalation`, which is part of the
`docs:all` task. In order for this to work a new constant
`RUN_AS_OPTIONS` which is array of escalation-related options was added
to the transport options module and is `concat`ed to relevant options
constants for the SSH and Local transport configuration classes.

!no-release-note

Closes #1867