(fun) Add rainbow output format
This adds a new output format option `rainbow` which behaves like the
human outputt except that it will colorize success messages with rainbow
colors instead of green if the system supports it. This works by adding
a new outputter class `Bolt::Outputter::Rainbow` that inherits from the
`Bolt::Outputter::Human` class and overrides the `colorize` and
`print_summary` functions. The colorize function in the rainbow class
will color messages labelled as `:green` or `:rainbow` with rainbow
colors using an algorithm copied from the [lolcat
project](https://github.com/busyloop/lolcat) (which is licensed for free
modification and reuse). The `print_summary` function is overridden to
colorize success messages that aren't colorized in the Human class.

This option can be configured on the cli with `--format rainbow` or in
config with `format: rainbow`.

This option is not available on Windows, and will throw a validation
error.

!feature

* **Add rainbow output format**

  This adds a new format option `rainbow` which will print success
  messages in rainbow colors. This option is not available on Windows.