(GH-1871) Don't infer project name from directory name
Previously, we would infer the project name from the directory name if
no name was specified in the bolt-project.yaml. This behavior proved to
be confusing for users, so we now require `name` to be specified
in bolt-project.yaml and otherwise will warn and not load project
content.

We previously recognized `<author>-<name>` directory names as valid
project names (using only the `<name>` portion) as that is a common
repo naming scheme for modules. As the project name must now be
specified explicitly, we instead require it be simply the name portion.

!deprecation

  * **Project names must be explicitly specified** ([#1871](https://github.com/puppetlabs/bolt/issues/1871))

    Proejct names must now be specified in `bolt-project.yaml` in order
    for project-level content to be loaded, rather than the name being
    inferred by the name of the project directory.