(GH-1904) Localhost default config overrides group-level config
Previously, the default config for 'localhost' was merged as part of the
group data collected for that target, making it effectively group-level
data. Because all targets are considered part of the 'all' group
top-level config would override localhost's default config, which made
it difficult to set defaults for all targets except localhost. This
modifies how we load localhost defaults to now merge them as part of
initializing the `Bolt::Inventory::Target` object, meaning that any time
a target with name 'localhost' is created (either via the inventory or
by `Target.new` in a plan) it will pick up the localhost defaults at the
target level. This means that now the only way to override the defaults
is to set that config at the target-level.

Closes #1904

!bug

* **Consider localhost default config target-level instead of group-level** ([1904](https://github.com/puppetlabs/bolt/issues/1904))

  Previously the 'localhost' special default config was merged at the
  group-level, meaning that group-level config in the inventory would
  override it. The config is now target-level, and must be overridden at
  the target-level in inventory.