(GH-1927) Fix powershell mkdir method
!bug

* **Fix use of spaces in powershell mkdir method** ([#1927](https://github.com/puppetlabs/bolt/issues/1927))

Fixes the powershell mkdir command to correctly handle paths with spaces in them. When passed to the command line, paths have to be quoted, the previous code did not handle this.

This uses double quotes instead of single quotes to allow string interpolation to happen when it is finally passed to PowerShell.