(GH-1899) Accept env_vars for commands and scripts
Previously, we had no way to specify environment variables when running
commands or scripts. There are many reasons a user might need to set
environment variables:

* To set up the execution context for a command
* To pass arguments to scripts that expect them as env vars
* To pass secrets to commands rather than passing them as args

We now support passing `_env_vars` to `run_command()` and `run_script()`
which will set the passed variables when running the command or script.

!feature

* **Set environment variables for commands and scripts** ([#1899](https://github.com/puppetlabs/bolt/issues/1899))

  The `run_command()` and `run_script()` plan functions now support an
  `_env_vars` argument which accepts a Hash of environment variable
  declarations to set when running the command/script.