(GH-1931) Add '--project' as an alias for '--boltdir'
This adds a `--project` CLI flag as an alias for `--boltdir`. If both
options are specified on the command line the rightmost option will be
used. `--project` behaves exactly like `--boltdir`. It also updates
documentation to use `--project` in place of `--boltdir`, since we
consistently refer to the concept as a 'project' in the documentation
and don't often use 'Boltdir'.

This additionally replaces `Boltdir` with `project` for all Bolt CLI
flag descriptions.

Closes #1931

!feature

* **Add --project as an alias for the --boltdir CLI flag** ([1931](https://github.com/puppetlabs/bolt/issues/1931))

  New CLI flag `--project` can be used in place of `--boltdir`.