(GH-1894) Do not load projects from world-writable directories
This updates the `Bolt::Project` class to raise an error if the
specified project directory is world-writable. This change is meant to
prevent malicious users from modifying project content that could alter
how Bolt runs and any secure information it might use.

This change only applies to Unix-like systems.

Users who wish to override this behavior and force Bolt to load a
project from a world-writable directory can set the `BOLT_PROJECT`
environment variable to the project directory path.

!feature

* **Do not load projects from world-writable directories**
  ([#1894](https://github.com/puppetlabs/bolt/issues/1894))

  Bolt now raises an error when attempting to load a project from a
  world-writable directory on Unix-like systems. Users who wish to
  override this behavior and run a project from a world-writable
  directory should can set the `BOLT_PROJECT` environment variable
  to the project directory path.