(GH-1920) Add --log-level CLI option
This adds a new `--log-level` CLI option that accepts a log level. When
using this option, any existing configuration for the console's log
level will be overridden with the specified level. The `--log-level`
option cannot be used in conjunction with the `--debug` option.

This also adds a warning that the `--debug` option is deprecated and
that Bolt will set `--log-level debug` instead.

!feature

* **Added `--log-level` CLI option**
  ([#1920](https://github.com/puppetlabs/bolt/issues/1920))

  The new `--log-level` CLI option can be used to override the console's
  log level. It accepts the following log levels: `debug`, `info`,
  `notice`, `warn`, `error`, `fatal`, `any`.