(GH-1928) Upload files to directories with the correct name
Previously, uploading a file to a directory on some transports would
cause the file to be renamed to match the name of the directory, ie.
`bolt file upload file.txt /dir` would result in a file called
`/dir/dir`. This happened because we create a tempfile that we move into
place, and we computed the name of the tempfile using the basename of
the destination, assuming it was always a file. We now use the basename
of the source file as the name of the tempfile, which preserves the name
when the destination is a directory and still causes the file to be
written with the propery destination name if the destination is a file.

!bug

* **Upload files with the correct name when destination is a directory** ([#1928](https://github.com/puppetlabs/bolt/issues/1928))

  Files uploaded to directories will now retain the name of the original
  source file rather than changing their name to the same name as the
  destination directory. This also fixes the case where the destination
  was `.`