(bug) Fix --transport bug
This fixes a bug introduced by #1923 that would not set the
default transport when using the `--transport` CLI option. Previously,
when normalizing the CLI overrides during config initialization, the
`OPTIONS` hash was used to slice out potential config options from the
CLI overrides hash. However, the `OPTIONS` hash does not include the
`transport` option, so the transport override would not be picked up.
This adds `transport` as an additional key to slice out of the overrides
hash.

!no-release-note