(GH-1938) Add warn_once logging utility
This adds a new class method called `warn_once` to the `Bolt::Logger`
class. `warn_once` keeps an instance variable of warning types that have
been issued and verifies that each warning is only sent to the logger
once. The warning uses it's own logger in the Bolt::Logger class, not
the logger from the calling class. The warning is also kept within a
mutex, so is only accessed by one thread at a time to prevent a warning
from being printed multiple times. This method is accessible from
anywhere in Bolt.

Closes #1938

!no-release-note