(bug-fix) Find embedded projects specified on the CLI
This modifies how projects specified with `--project` or `--boltdir` are
found. Previously the provided path would be used as the path to the
project, where bolt configuration was expected to be found. However with
embedded projects - those that store files in `Boltdir` - users would
specify the parent directory of `Boltdir` expect the project to be
loaded as if they were executing Bolt from that parent directory. This
modifies how we handle project directories first looks for a `Boltdir`
in the specified directory, and if none is present use the specified
directory as the project root.

!bug

* **Load projects with an embedded Boltdir when specified on the CLI**

  We now look for a `Boltdir` in the directory specified by `--project`
  or `--boltdir` and use that if present. If not, we use the specified
  path as the project root.