(GH-1932) Add location of plan failures in error messages
This updates the error raised by a plan failure to include the location
of the plan failure, including the filepath, line, and column of the
error's source. This information is included in both the error message
and under the `details` key of the error hash. Previously, this
information was discarded by Bolt when rescuing an error raised by
Puppet, resulting in only the cause of the error to be included in the
message.

!feature

* **Add location of plan failures in error messages**
  ([#1923](https://github.com/puppetlabs/bolt/issues/1923))

  Errors raised by plan failures now include the location of the plan
  failure, including the filepath, line, and column.