(packaging) Make the paint gem required by the rainbow outputter optional
Given we dont want to ship the paint gem with PE this commit makes the gem optional. The gem will only attempt to be loaded when the rainbow outputter is requested. The paint gem will ship with bolt packages and is included in the Gemfile, but is ommited from the gemspec.

!no-release-note