(GH-1900) Generate JSON schemas with plugin references
This contains changes that enable generating the JSON schemas using a
rake task using data contained in the `Bolt::Config::Options` and
`Bolt::Config::Transport::TRANSPORT_OPTIONS` constants.

A new Rakefile `schemas/Rakefile` contains rake tasks that generate
each of the JSON schemas. These tasks can be run using the following
command:

```shell
$ rake -f schemas/Rakefile schemas
```

This also updates the `Bolt::Config::OPTIONS` and
`Bolt::Config::Transport::TRANSPORT_OPTIONS` constants to include
additional data for each option that is included in the JSON schemas.
The additional keys are direct equivalents to keys recognized by the
JSON Schema standard, Draft 07. The constants also include several
metadata keys such as `:_example` and `:_default`, which are used to
generate docuemntation for the configuration files.

This change also updates the JSON schemas to support plugin references.
The `:_plugin` metadata key can be set on an option to indicate that the
JSON schema should include a plugin reference as a valid value for the
option.

!bug

* **Support plugin references in JSON schemas**
  ([#1900](https://github.com/puppetlabs/bolt/issues/1900))

  The JSON schemas no longer mark plugin references as invalid values
  when the option can accept a plugin reference. Previously, the schemas
  would mark any plugin reference as an invalid value.