(bug) Validate 'inventory-config' option in bolt-defaults.yaml
This adds some validation checks for the `inventory-config` option in
the `bolt-defaults.yaml` file. Previously, setting this option to a
value other than a hash would result in an unhelpful error, as Bolt was
assuming that the value would always be a hash. Additionally, if the
value was a plugin reference, the plugin reference would be merged with
the top-level config, potentially resulting in config values being
overwritten and resulting in the entire config hash being considered a
plugin reference.

Now, Bolt will first check that the `inventory-config` option is a hash
and is not a plugin reference before continuing with loading config.

!bug

* **Validate `inventory-config` option in `bolt-defaults.yaml`**

  Bolt now checks that the `inventory-config` option in a
  `bolt-defaults.yaml` file is a hash and is not a plugin reference
  before merging configuration files. Previously, setting this value to
  a hash or plugin reference would raise an unhelpful error.