(GH-1895) Add action to trigger updating ps module in bolt-vanagon
In order to ship the powershell cmdlets module in the Bolt MSI package
we need generate the module from it's constituent files and commit it to
bolt-vanagon. This will happen by creating a repository-dispatch event
whenever changes to the powershell files are committed to master. The
repository-dispatch event sends a payload to the specified repository
which can be used to trigger another Github Action. This
repository-dispatch will trigger an action that generates the powershell
module and commits it to master of bolt-vanagon. This should ensure that
the module in bolt-vanagon stays up to date with the module components
in bolt.

!no-release-note