(GH-1917) Don't create Bolt Config object before loading config
Previously when we created a new instance of the `Bolt::CLI` object we
also loaded the default `Bolt::Config` object in order to create a
default outputter to print errors with. This caused any errors loading
config or the default project to be unhandled and backtrace. The error
was especially annoying because it used the default project path at
`~/.puppetlabs/bolt`, regardless of whether that was the actual project
that would be used during the Bolt run.

This commit moves the same behavior of the default outputter directly
into the CLI (namely printing any error messages to stdout, in red if
it's a tty) so that we no longer need to load the default project for
every invocation of Bolt. This makes it so any errors that occur when
loading config are printed, and their exit code returned, similar to
other errors raised in Bolt::CLI.

Closes #1917

!bug

* **Don't load default project for every Bolt invocation** ([1917](https://github.com/puppetlabs/bolt/issues/1917)

  We no longer load the default project at `~/.puppetlabs/bolt` for
  every Bolt invocation. When projects are loaded, the errors are now
  handled and don't backtrace.