(GH-1967) Detect puppet agent install path on windows
Correctly detects the Puppet Agent install path on Windows during the initialization steps in a WinRM connection. It then uses the correct install path to amend the environment PATH variable.

The error text was modified to more clearly state we looked in known locations but could not find a valid install path.

!bug

* **Correctly detect puppet agent install path on windows** ([#1967](https://github.com/puppetlabs/bolt/issues/1967))

Correctly detects the Puppet Agent install path on Windows during the initialization steps in a WinRM connection