(GH-1868) Add support for downloading files from targets
This adds support for a `bolt file download` command that downloads a
file or directory from a target to a destination directory on the local
system. File downloads are supported on all transports _except for
`pcp`_.

The `bolt file download` command expects two arguments:

- **source:** The path to the file or directory to be downloaded from
the target
- **destination:** The path to the destination directory on the local
system

The destination directory accepts either an absolute path or a relative
path. If a relative path is given, it expands relative to the project's
`downloads` directory. For example, the default project's downloads
directory is located at `~/.puppetlabs/bolt/downloads/`.

When downloading content, Bolt creates a directory for each target
within the destination directory. The directory for each target has a
name matching the URL-encoded safe name for the target. All content
downloaded for the target is stored in this directory.

!feature

* **Added `bolt file download` CLI command**
  ([#1868](https://github.com/puppetlabs/bolt/issues/1868))

  Users can now download files and directories from targets to the local
  system using the `bolt file download` CLI command. This command
  accepts a path to the file or directory to download from the targets
  and a path to a destination directory on the local system. The
  destination directory is expanded relative to the project downloads
  directory (i.e. `<project>/downloads/`).