(GH-1868) Add `download_file` plan function
This adds a new `download_file` plan function which can be used to
download a specified file or directory from a list of targets to the
local system.

Similar to the `bolt file download` CLI command, this plan function
accepts `source` and `destination` arguments. The `source` argument is
the path to the file or directory to download from the targets, while
the `destination` argument is a path to a directory on the local system
that the files should be saved to.

The path to the destination directory is expanded relative to the
project's `downloads` directory (i.e. `<project>/downloads/`).

The result for this plan function includes the `path` to the downloaded
file on the local system.

!feature

* **Add `download_file` plan function**

  The `download_file` plan function can be used to download a file or
  directory from a list of targets to a destination directory on the
  local system. The result returned from this function includes the path
  to the downloaded file on the local system.