(GH-1868) Add YAML plan download step
This adds a new YAML plan step for Bolt's file download functionality.
The step uses the following keys:

- `download`: The path to the file or directory to download from the
targets
- `destination`: The path to the destination directory on the local
system
- `targets`: The list of targets to download from

This also updates the YAML plan upload step to use an `upload` key in
place of a `source` key. Previously, the upload step used a `source`
key to identify the step as an upload step. This ambiguous key would
have been confusing with the addition of a download step, as it also
expects a source path.

The `source` key for YAML plan upload steps has been deprecated. Bolt
will raise a helpful warning when a plan step includes a `source` key.

!feature

* **Add YAML plan download step**
  ([#1868](https://github.com/puppetlabs/bolt/issues/1868))

  YAML plans now support a download step which can be used to download a
  file or directory from a list of targets to a destination directory on
  the local system.

!deprecation

* **Deprecate `source` key for YAML plan upload step**
  ([#1868](https://github.com/puppetlabs/bolt/issues/1868))

  The `source` key used in YAML plan upload steps has been deprecated in
  favor of a less ambiguous `upload` key.