(maint) Support Rainbow format on windows
The Paint gem detects the OS and what terminal it is running in and makes some decisions about what color scheme is supported. On windows it assumes no color is supported unless the ANSICON or ConEmuANSI enviroment variables are set. Until that gem is updated to detect Windows 10 and treat it as valid, we can force the Paint gem to work by setting the mode manually. Setting `Paint.mode` to TRUE_COLOR (0xFFFFFF) if we are on Windows forces the Paint gem to color the string output. This works in PowerShell 5.1, 6, 7, WSL, and even in cmd.exe on Windows 10.

Reference https://github.com/janlelis/paint/blob/4492d100d885d5c2d46305417738b75e69db7f03/lib/paint.rb#L177-L194

!feature

  * **Support Rainbow format on windows**

    Set `Paint.mode` to `TRUE_COLOR` to enable ansi color sequences on Windows