(maint) Don't expose Pathname object to Puppet
Previously, when passing the Bolt project to Puppet to turn into a
module, we were exposing the "path" variable as a Pathname instance.
Puppet expects the path will always be a string and fails in at least
one case if it receives another value. Since we just use Pathname for
our own internal tracking, we now turn it into a string in the `to_h`
method for Puppet's consumption.

!bug

  * **Don't fail for tasks referring to project-level files**
    ([#1984](https://github.com/puppetlabs/bolt/issues/1984))

    Tasks with a `files` key that refers to files that exist at the
    project level will now load properly rather than throwing an error.