(GH-1994) Configure timeouts in PuppetDB client
This adds support for a `connect_timeout` and `read_timeout` option
in `puppetdb` config. The `connect_timeout` value is used to set the
connection timeout for the HTTP client that connects to PuppetDB, while
`read_timeout` is used to set the receive timeout for the HTTP client.
The value for each of these options must be a positive integer. If the
options are not set, both are set to the default values defined by the
`httpclient` Ruby library (currently 60 seconds).

!feature

* **Configure connection and read timeout length in PuppetDB client**
  ([#1994](https://github.com/puppetlabs/bolt/issues/1994))

  Users can now configure the connection and read timeout length for the
  PuppetDB client with the `connect_timeout` and `read_timeout`
  options under the `puppetdb` config option.