(GH-1998) Create rakelib directory for Rake tasks
This adds a new `rakelib` directory in the repository root. This
directory contains rake tasks for generating docs, schemas, powershell
module files, and running tests. Each of these files is automatically
loaded by rake when running the `rake` command.

This also wraps several tasks and `require` statements in a `begin`
block, allowing rake to attempt to load development gems and still
continue if they are not available in the environment. This lets other
tasks, such as the powershell module tasks, to be loaded and executed
even if other tasks cannot be loaded.

Several tasks have also been removed, such as `rubocop` and
`update_submodules`, as they were either redundant or unused.

Lastly, this renames several tasks to better indicate what the tasks are
used for. All of the tasks previously under the `integration` namespace
are now under the `tests` namespace. Standalone rspec tasks, such as
`unit`, have also been placed under the `tests` namespace. Tasks used in
CI have been placed under the `ci` namespace. Tasks under the `ci`
namespace have also had their descriptions removed so they are not
displayed in the list of tasks when running `rake -T`, as they typically
will not be run directly by users.

!no-release-note