(bug) Do not re-rescue errors raised before config is loaded
This fixes a bug introduced in #1917 that caused errors raised before
config is loaded to be re-raised and rescued by `Bolt::CLI#parse` and
sent to the outputter. This resulted in a `NoMethodError` being raised,
as the outputter relies on the config being loaded when it is
instantiated.

This seperates the logic that was in the `#parse` method into three
separate methods. The `#parse_command` method handles parsing and
validating the command, the `#load_config` method loads the project and
config, and the `#finalize_setup` method finishes configuring Bolt and
issuing any warnings from the previous steps.

Splitting the logic into separate methods allows us to handle errors
raised before and after config is loaded differently.

!bug

* **Do not re-rescue errors raised before config is loaded**

  Errors raised before Bolt's configuration was loaded were raising a
  second, ambiguous `NoMethodError`. Bolt now correctly handles errors
  raised before configuration is loaded and will no longer trigger
  additional errors.